TIP-482: move InvalidItemFromViolationsException
